Smell and refactor analysis moved to new class;
